chore: rename mappor to ippo
