+++ examples/jax/debugging/simple_spread/feedforward/decentralised/run_ippo.py
@@ -13,31 +13,31 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Example running MAPPO on debug MPE environments."""
+"""Example running IPPO on debug MPE environments."""
 import functools
+import os
 from datetime import datetime
 from typing import Any
 
-import numpy as np
 import optax
 from absl import app, flags
-from acme.jax.networks.atari import DeepAtariTorso
-from supersuit import dtype_v0
 
-from mava.systems.jax import mappo
-from mava.utils.environments import pettingzoo_utils
+from mava.systems.jax import ippo
+from mava.utils.environments import debugging_utils
 from mava.utils.loggers import logger_utils
 
+# Without this flag, JAX uses the whole GPU from the beginning and our trainer crashes.
+os.environ["XLA_PYTHON_CLIENT_PREALLOCATE"] = "false"
 FLAGS = flags.FLAGS
 flags.DEFINE_string(
-    "env_class",
-    "butterfly",
-    "Pettingzoo environment class, e.g. atari (str).",
+    "env_name",
+    "simple_spread",
+    "Debugging environment name (str).",
 )
 flags.DEFINE_string(
-    "env_name",
-    "cooperative_pong_v5",
-    "Pettingzoo environment name, e.g. pong (str).",
+    "action_space",
+    "discrete",
+    "Environment action space type (str).",
 )
 
 flags.DEFINE_string(
@@ -45,7 +45,6 @@ flags.DEFINE_string(
     str(datetime.now()),
     "Experiment identifier that can be used to continue experiments.",
 )
-
 flags.DEFINE_string("base_dir", "~/mava", "Base dir to store experiments.")
 
 
@@ -57,19 +56,16 @@ def main(_: Any) -> None:
     """
     # Environment.
     environment_factory = functools.partial(
-        pettingzoo_utils.make_environment,
-        env_class=FLAGS.env_class,
+        debugging_utils.make_environment,
         env_name=FLAGS.env_name,
-        env_preprocess_wrappers=[(dtype_v0, {"dtype": np.float32})],
+        action_space=FLAGS.action_space,
     )
 
     # Networks.
     def network_factory(*args: Any, **kwargs: Any) -> Any:
-        obs_net_forward = lambda x: DeepAtariTorso()(x)  # noqa: E731
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(64,),
-            critic_layer_sizes=(256,),
-            observation_network=obs_net_forward,
+        return ippo.make_default_networks(  # type: ignore
+            policy_layer_sizes=(254, 254, 254),
+            critic_layer_sizes=(512, 512, 256),
             *args,
             **kwargs,
         )
@@ -94,7 +90,7 @@ def main(_: Any) -> None:
     )
 
     # Create the system.
-    system = mappo.MAPPOSystem()
+    system = ippo.IPPOSystem()
 
     # Build the system.
     system.build(
@@ -108,6 +104,7 @@ def main(_: Any) -> None:
         num_epochs=15,
         num_executors=1,
         multi_process=True,
+        clip_value=False,
     )
 
     # Launch the system.
