+++ examples/jax/debugging/simple_spread/feedforward/decentralised/run_ippo_single_process.py
@@ -12,20 +12,34 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Example running feedforward MADQN on Flatland."""
 
+"""Example running IPPO on debug MPE environments."""
 import functools
+import os
 from datetime import datetime
-from typing import Any, Dict
+from typing import Any
 
+import launchpad as lp
 import optax
 from absl import app, flags
 
-from mava.systems.jax import mappo
-from mava.utils.environments.flatland_utils import make_environment
+from mava.systems.jax import ippo
+from mava.utils.environments import debugging_utils
 from mava.utils.loggers import logger_utils
 
+# Without this flag, JAX uses the whole GPU from the beginning and our trainer crashes.
+os.environ["XLA_PYTHON_CLIENT_PREALLOCATE"] = "false"
 FLAGS = flags.FLAGS
+flags.DEFINE_string(
+    "env_name",
+    "simple_spread",
+    "Debugging environment name (str).",
+)
+flags.DEFINE_string(
+    "action_space",
+    "discrete",
+    "Environment action space type (str).",
+)
 
 flags.DEFINE_string(
     "mava_id",
@@ -35,44 +49,26 @@ flags.DEFINE_string(
 flags.DEFINE_string("base_dir", "~/mava", "Base dir to store experiments.")
 
 
-# flatland environment config
-flatland_env_config: Dict = {
-    "n_agents": 3,
-    "x_dim": 30,
-    "y_dim": 30,
-    "n_cities": 2,
-    "max_rails_between_cities": 2,
-    "max_rails_in_city": 3,
-    "seed": 0,
-    "malfunction_rate": 1 / 200,
-    "malfunction_min_duration": 20,
-    "malfunction_max_duration": 50,
-    "observation_max_path_depth": 30,
-    "observation_tree_depth": 2,
-}
-
-
 def main(_: Any) -> None:
-    """Run main script
+    """Main script for running system."""
 
-    Args:
-        _ : _
-    """
+    system = ippo.IPPOSystem()
 
-    # WARNING (dries): This code has not been run yet. There might still
-    # be runtime errors in the code.
+    # Environment.
+    environment_factory = functools.partial(
+        debugging_utils.make_environment,
+        env_name="simple_spread",
+        action_space="discrete",
+    )
 
     # Networks.
     def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
+        return ippo.make_default_networks(  # type: ignore
             policy_layer_sizes=(254, 254, 254),
             critic_layer_sizes=(512, 512, 256),
             *args,
             **kwargs,
         )
-
-    # Environment.
-    environment_factory = functools.partial(make_environment, **flatland_env_config)
 
     # Used for checkpoints, tensorboard logging and env monitoring
     experiment_path = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
@@ -87,30 +83,25 @@ def main(_: Any) -> None:
         time_stamp=FLAGS.mava_id,
         time_delta=log_every,
     )
-
     # Optimizer.
     optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-1e-4)
+        optax.clip_by_global_norm(40.0),
+        optax.adam(1e-4),
     )
-
-    # Create the system.
-    system = mappo.MAPPOSystem()
-
-    # Build the system.
+    # Build the system
     system.build(
         environment_factory=environment_factory,
         network_factory=network_factory,
         logger_factory=logger_factory,
         experiment_path=experiment_path,
         optimizer=optimizer,
+        multi_process=False,
         run_evaluator=True,
-        sample_batch_size=5,
-        num_epochs=15,
         num_executors=1,
-        multi_process=True,
+        max_queue_size=500,
+        sample_batch_size=5,
+        lp_launch_type=lp.LaunchType.LOCAL_MULTI_PROCESSING,
     )
-
-    # Launch the system.
     system.launch()
