+++ examples/jax/petting_zoo/butterfly/cooperative_pong/feedforward/decentralised/run_ippo.py
@@ -13,28 +13,31 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Example running MAPPO on debug MPE environments."""
+"""Example running IPPO on debug MPE environments."""
 import functools
 from datetime import datetime
 from typing import Any
 
+import numpy as np
 import optax
 from absl import app, flags
+from acme.jax.networks.atari import DeepAtariTorso
+from supersuit import dtype_v0
 
-from mava.systems.jax import mappo
-from mava.utils.environments import debugging_utils
+from mava.systems.jax import ippo
+from mava.utils.environments import pettingzoo_utils
 from mava.utils.loggers import logger_utils
 
 FLAGS = flags.FLAGS
 flags.DEFINE_string(
-    "env_name",
-    "simple_spread",
-    "Debugging environment name (str).",
+    "env_class",
+    "butterfly",
+    "Pettingzoo environment class, e.g. atari (str).",
 )
 flags.DEFINE_string(
-    "action_space",
-    "discrete",
-    "Environment action space type (str).",
+    "env_name",
+    "cooperative_pong_v5",
+    "Pettingzoo environment name, e.g. pong (str).",
 )
 
 flags.DEFINE_string(
@@ -42,6 +45,7 @@ flags.DEFINE_string(
     str(datetime.now()),
     "Experiment identifier that can be used to continue experiments.",
 )
+
 flags.DEFINE_string("base_dir", "~/mava", "Base dir to store experiments.")
 
 
@@ -53,23 +57,25 @@ def main(_: Any) -> None:
     """
     # Environment.
     environment_factory = functools.partial(
-        debugging_utils.make_environment,
+        pettingzoo_utils.make_environment,
+        env_class=FLAGS.env_class,
         env_name=FLAGS.env_name,
-        action_space=FLAGS.action_space,
+        env_preprocess_wrappers=[(dtype_v0, {"dtype": np.float32})],
     )
 
     # Networks.
     def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
-            policy_layer_sizes=(254, 254, 254),
-            critic_layer_sizes=(512, 512, 256),
-            single_network=False,
+        obs_net_forward = lambda x: DeepAtariTorso()(x)  # noqa: E731
+        return ippo.make_default_networks(  # type: ignore
+            policy_layer_sizes=(64,),
+            critic_layer_sizes=(256,),
+            observation_network=obs_net_forward,
             *args,
             **kwargs,
         )
 
-    # Checkpointer appends "Checkpoints" to checkpoint_dir
-    checkpoint_subpath = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
+    # Used for checkpoints, tensorboard logging and env monitoring
+    experiment_path = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
 
     # Log every [log_every] seconds.
     log_every = 10
@@ -83,25 +89,20 @@ def main(_: Any) -> None:
     )
 
     # Optimizer.
-    policy_optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-1e-4)
-    )
-
-    critic_optimizer = optax.chain(
+    optimizer = optax.chain(
         optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-1e-4)
     )
 
     # Create the system.
-    system = mappo.MAPPOSystemSeparateNetworks()
+    system = ippo.IPPOSystem()
 
     # Build the system.
     system.build(
         environment_factory=environment_factory,
         network_factory=network_factory,
         logger_factory=logger_factory,
-        experiment_path=checkpoint_subpath,
-        policy_optimizer=policy_optimizer,
-        critic_optimizer=critic_optimizer,
+        experiment_path=experiment_path,
+        optimizer=optimizer,
         run_evaluator=True,
         sample_batch_size=5,
         num_epochs=15,
