+++ examples/jax/smac/feedforward/decentralised/run_ippo_sep_networks.py
@@ -12,33 +12,25 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Run feedforward MADQN on SMAC."""
 
-"""Example running MAPPO on debug MPE environments."""
+
 import functools
-import os
 from datetime import datetime
 from typing import Any
 
-import launchpad as lp
 import optax
 from absl import app, flags
 
-from mava.systems.jax import mappo
-from mava.utils.environments import debugging_utils
+from mava.systems.jax import ippo
+from mava.utils.environments.smac_utils import make_environment
 from mava.utils.loggers import logger_utils
 
-# Without this flag, JAX uses the whole GPU from the beginning and our trainer crashes.
-os.environ["XLA_PYTHON_CLIENT_PREALLOCATE"] = "false"
 FLAGS = flags.FLAGS
 flags.DEFINE_string(
-    "env_name",
-    "simple_spread",
-    "Debugging environment name (str).",
-)
-flags.DEFINE_string(
-    "action_space",
-    "discrete",
-    "Environment action space type (str).",
+    "map_name",
+    "3m",
+    "Starcraft 2 micromanagement map name (str).",
 )
 
 flags.DEFINE_string(
@@ -50,28 +42,26 @@ flags.DEFINE_string("base_dir", "~/mava", "Base dir to store experiments.")
 
 
 def main(_: Any) -> None:
-    """Main script for running system."""
+    """Example running feedforward MADQN on SMAC environment."""
 
-    system = mappo.MAPPOSystem()
+    # WARNING (dries): This code has not been run yet. There might still
+    # be runtime errors in the code.
 
-    # Environment.
-    environment_factory = functools.partial(
-        debugging_utils.make_environment,
-        env_name="simple_spread",
-        action_space="discrete",
-    )
+    # Environment
+    environment_factory = functools.partial(make_environment, map_name=FLAGS.map_name)
 
     # Networks.
     def network_factory(*args: Any, **kwargs: Any) -> Any:
-        return mappo.make_default_networks(  # type: ignore
+        return ippo.make_default_networks(  # type: ignore
             policy_layer_sizes=(254, 254, 254),
             critic_layer_sizes=(512, 512, 256),
+            single_network=False,
             *args,
             **kwargs,
         )
 
-    # Used for checkpoints, tensorboard logging and env monitoring
-    experiment_path = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
+    # Checkpointer appends "Checkpoints" to checkpoint_dir
+    checkpoint_subpath = f"{FLAGS.base_dir}/{FLAGS.mava_id}"
 
     # Log every [log_every] seconds.
     log_every = 10
@@ -83,25 +73,35 @@ def main(_: Any) -> None:
         time_stamp=FLAGS.mava_id,
         time_delta=log_every,
     )
+
     # Optimizer.
-    optimizer = optax.chain(
-        optax.clip_by_global_norm(40.0),
-        optax.adam(1e-4),
+    policy_optimizer = optax.chain(
+        optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-3e-4)
     )
-    # Build the system
+    critic_optimizer = optax.chain(
+        optax.clip_by_global_norm(40.0), optax.scale_by_adam(), optax.scale(-1e-3)
+    )
+
+    # Create the system.
+    system = ippo.IPPOSystemSeparateNetworks()
+
+    # Build the system.
     system.build(
         environment_factory=environment_factory,
         network_factory=network_factory,
         logger_factory=logger_factory,
-        experiment_path=experiment_path,
-        optimizer=optimizer,
-        multi_process=False,
+        experiment_path=checkpoint_subpath,
+        policy_optimizer=policy_optimizer,
+        critic_optimizer=critic_optimizer,
         run_evaluator=True,
-        num_executors=1,
-        max_queue_size=500,
         sample_batch_size=5,
-        lp_launch_type=lp.LaunchType.LOCAL_MULTI_PROCESSING,
+        num_epochs=15,
+        num_executors=1,
+        multi_process=True,
+        clip_value=True,
     )
+
+    # Launch the system.
     system.launch()
