+++ mava/systems/jax/ippo/README.md
@@ -1,0 +1,10 @@
+# Independent Proximal Policy Optimization (IPPO)
+
+This system is an extension of the Proximal Policy Optimization (PPO) ([Schulman et al., 2017]) algorithm for the multi-agent setting. The system is currently trained in a decentralised manner and is hence termed IPPO. Centralised training, termed MAPPO, will be realeased shortly. For recent experimental results concerning Multi-agent PPO, see the following paper by [Yu et al. (2021)].
+
+<p style="text-align:center;">
+<img src="https://raw.githubusercontent.com/instadeepai/Mava/develop/docs/images/ppo.png" width="60%">
+</p>
+
+[Schulman et al., 2017]: https://arxiv.org/pdf/1707.06347
+[Yu et al. (2021)]: https://arxiv.org/pdf/2103.01955
