+++ mava/systems/jax/ippo/__init__.py
@@ -13,13 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Default hyperparameters for MAPPO system."""
-from dataclasses import dataclass
-
-
-@dataclass
-class MAPPODefaultConfig:
-    sample_batch_size: int = 512
-    sequence_length: int = 20
-    period: int = 10
-    use_next_extras: bool = False
+"""Jax IPPO system."""
+from mava.systems.jax.ippo.networks import make_default_networks
+from mava.systems.jax.ippo.system import IPPOSystem, IPPOSystemSeparateNetworks
