+++ mava/systems/jax/ippo/networks.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Jax MAPPO system networks."""
+"""Jax IPPO system networks."""
 import dataclasses
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
