--- mava/systems/jax/mappo/README.md
@@ -1,10 +1,0 @@
-# Multi-Agent Proximal Policy Optimization (MAPPO)
-
-This system is an extension of the Proximal Policy Optimization (PPO) ([Schulman et al., 2017]) algorithm to the multi-agent setting. The system can be trained in a decentralised or centralised manner using a shared critic network. For recent experimental results concerning MAPPO, see the follow paper by [Yu et al. (2021)].
-
-<p style="text-align:center;">
-<img src="https://raw.githubusercontent.com/instadeepai/Mava/develop/docs/images/ppo.png" width="60%">
-</p>
-
-[Schulman et al., 2017]: https://arxiv.org/pdf/1707.06347
-[Yu et al. (2021)]: https://arxiv.org/pdf/2103.01955
